- hpcrun [csprof] and hpcprof [xcsprof]: Allow metrics to be either of integral or real type. - hpcprof [xcsprof]: Calling context tree nodes contain a link to metric descriptor information to interpret metrics correctly.

+++ trunk/src/hpctoolkit/csprof/lush/lushi.h
@@ -130,7 +130,7 @@ LUSHI_DECL(int, LUSHI_lip_write, ());
 // --------------------------------------------------------------------------
 
 LUSHI_DECL(int, LUSHI_has_concurrency, ());
-LUSHI_DECL(uint, LUSHI_get_concurrency, ());
+LUSHI_DECL(double, LUSHI_get_concurrency, ());
 
 
 #ifdef __cplusplus
